pyproject.toml
[tool.poetry]
name = "fuelpricesgr"
version = "0.1.0"
description = ""
authors = ["Kostas Kokkoros <mavroprovato@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.11.0"
boto3 = "^1.24.95"
fastapi = "^0.85.1"
fastapi-cache2 = {extras = ["redis"], version = "^0.1.9"}
gunicorn = "^20.1.0"
PyPDF2 = "^2.10.0"
python-dateutil = "^2.8.0"
python-dotenv = "^0.21.0"
requests = "^2.28.0"
tortoise-orm = "^0.19.2"
uvicorn = {extras = ["standard"], version = "^0.19.0"}

[tool.poetry.dev-dependencies]
coverage = "^6.4.4"
fabric = "^2.7.1"
patchwork = "^1.0.1"
pytest = "^7.1"
pylint = "^2.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
